+++ routes/users.js
@@ -1,38 +1,15 @@
 const express = require('express');
-const User = require('../models/user');
 const { isNotLoggedIn } = require('../middlewares/login-auth');
+const userController = require('../controllers/user-controller');
 
 const router = express.Router();
 
-router.get('/', isNotLoggedIn, (req, res, next) => {
-	const flashMessage = req.flash();
-	let message = '';
-
-	if (flashMessage.message) {
-		message = flashMessage.message[0];
-	}
-
-	res.render('sign-up', { title: 'Sign Up', message });
+router.get('/', isNotLoggedIn, (req, res) => {
+	userController.getSignUpPage(req, res);
 });
 
 router.post('/', isNotLoggedIn, async (req, res, next) => {
-	const { user_id, user_password } = req.body;
-
-	try {
-		const duplicatedId = await User.checkDuplicatedId(user_id, 'local');
-
-		if (duplicatedId) {
-			req.flash('message', '이미 사용중인 아이디입니다.');
-			return res.redirect('/users');
-		}
-
-		const signUpForm = { user_id, user_password, auth_provider: 'local' };
-		await User.signUp(signUpForm);
-
-		return res.redirect('/auth');
-	} catch (error) {
-		console.error(error);
-		next(error);
-	}
+	userController.signUp(req, res, next);
 });
+
 module.exports = router;
